analyzer: Move static functions used by `Npm` and `Yarn2` to `NpmSupport`
To compensate for the lost class namespacing, add "Npm" infixes to the
function names which also aligns with the existing functions in the
file.
